% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandPeakDataCols.R
\name{expandPeakDataCols}
\alias{expandPeakDataCols}
\title{Detect Local Minima and Maxima in a Density Function}
\usage{
expandPeakDataCols(methylModesResults)
}
\arguments{
\item{methylModesResults}{A data.table returned by methylModesBatch()}
}
\value{
A modified version of the input data table with these changes:
\enumerate{
\item Within each row, columns peakLocations, proportionSample, and
peakVariance (which are objects of class "list") are sorted in descending
order by proportionSample
\item Columns are appended to the end of the data frame. For peakLocations,
proportionSample, and peakVariance, the values listed within these columns
are split across new columns containing one value for each peak. The
number of new columns is equal to 3 * the maximum number of peaks found for
any probe. These columns are named with the following convention:
peakLocation_i, proportionSample_i, peakVariance_i, where i ranges from 1
to the maximum number of peaks
}
}
\description{
This function identifies the indices of local minima and maxima in a density function,
typically used for peak detection in DNA methylation data. It handles edge cases and
ensures proper boundary conditions for the density estimation.
}
\examples{
# Obtain MethylModes results from the toy dataset
# Note that this will take a minute or so
test <- methylModesBatch(methylModes::toyMultimodalData)
testExpanded <- expandPeakDataCols(test)
# Sort by second-largest peak as a measure of confidence in multimodality
testExpanded <- data.table::setorder(testExpanded, -proportionSample_2, na.last = TRUE)

}
