% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaManifestEPIC}
\alias{IlluminaManifestEPIC}
\title{Illumina EPIC Manifest Data}
\format{
A data.table with 865,919 rows and 7 columns:
\describe{
\item{IlmnID}{Probe ID (e.g., "cg00000029")}
\item{CHR}{Chromosome number}
\item{Relation_to_UCSC_CpG_Island}{Relation to CpG island (Island, N_Shore, S_Shore, etc.)}
\item{MAPINFO}{Base pair position}
\item{SNP_ID}{RSID of SNPs under the probe}
\item{SNP_DISTANCE}{Distance from CpG to SNPs under the probe}
\item{SNP_MinorAlleleFrequency}{MAF of SNPs under the probe}
}
}
\source{
Illumina HumanMethylationEPIC BeadChip annotation
}
\usage{
IlluminaManifestEPIC
}
\description{
Annotation data for Illumina HumanMethylationEPIC BeadChip array (v1.0).
Contains probe information including chromosome, location on chromosome,
position relative to CpG islands, and genetic variants under the probe.
Used in methylModes Shiny app to allow analysis on a subset of data.
Original manifest downloaded from \href{https://support.illumina.com/downloads/infinium-methylationepic-v1-0-product-files.html}{Illumina Support}.
}
\keyword{datasets}
