% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaManifest450k}
\alias{IlluminaManifest450k}
\title{Illumina 450k Manifest Data}
\format{
A data.table with 485,577 rows and 6 columns:
\describe{
\item{IlmnID}{Probe ID (e.g., "cg00000029")}
\item{CHR}{Chromosome number}
\item{Relation_to_UCSC_CpG_Island}{Relation to CpG island (Island, N_Shore, S_Shore, etc.)}
\item{MAPINFO}{Base pair position}
\item{Probe_SNPs_10}{SNPs within 10bp of probe}
\item{Probe_SNPs}{SNPs 10-50bp from probe}
}
}
\source{
Illumina HumanMethylation450 BeadChip annotation
}
\usage{
IlluminaManifest450k
}
\description{
Annotation data for Illumina HumanMethylation450 BeadChip array (v1.2).
Contains probe information including chromosome, location on chromosome,
position relative to CpG islands, and genetic variants under the probe.
Used in methylModes Shiny app to allow analysis on a subset of data.
Original manifest downloaded from \href{https://support.illumina.com/downloads/infinium_humanmethylation450_product_files.html}{Illumina Support}
}
\keyword{datasets}
