% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IlluminaManifestEPICv2}
\alias{IlluminaManifestEPICv2}
\title{Illumina EPIC v2 Manifest Data}
\format{
A data.table with 926,371 rows and 7 columns:
\describe{
\item{IlmnID}{Probe ID (e.g., "cg25324105")}
\item{CHR}{Chromosome number}
\item{Relation_to_UCSC_CpG_Island}{Relation to nearest CpG island (Island, N_Shore, S_Shore, etc.)}
\item{MAPINFO}{Base pair position}
\item{SNP_ID}{RSID of SNPs under the probe}
}
}
\source{
Illumina HumanMethylationEPIC v2 BeadChip annotation
}
\usage{
IlluminaManifestEPICv2
}
\description{
Annotation data for Illumina HumanMethylationEPIC BeadChip array (v2.0).
Contains probe information including chromosome, location on chromosome,
position relative to CpG islands, and genetic variants under the probe.
Used in methylModes Shiny app to allow analysis on a subset of data.
SNP_DISTANCE and SNP_MinorAlleleFrequency are ommited in this copy
of the manifest due to file size constraints.
Original manifest downloaded from \href{https://support.illumina.com/downloads/infinium-methylationepic-v2-0-product-files.html}{Illumina Support}.
}
\keyword{datasets}
