\name{methylModes-package}
\alias{methylModes-package}
\alias{methylModes}
\docType{package}
\title{
\packageTitle{methylModes}
}
\description{
\packageDescription{methylModes}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{methylModes}

Maintainer: \packageMaintainer{methylModes}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = methylModes)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{methylModes}}
%\section{Documentation Index}{\packageIndices{methylModes}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
