% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyMultimodalData}
\alias{toyMultimodalData}
\title{Toy Data for methylModes}
\format{
A matrix with 10000 rows and 800 columns:
\describe{
\item{rows}{Probe IDs (first 10000 from IlluminaManifest450k IlmnID column)}
\item{columns}{Sample IDs (subject1 through subject800)}
\item{values}{Beta values between 0 and 1}
}
}
\source{
Generated using \code{inst/generateToyMultimodalData.R}
}
\usage{
toyMultimodalData
}
\description{
A dataset containing randomly generated multimodal distributions.
Rather than being designed to mimic DNA methylation betas distributions,
this data is intended to verify functionality of the Shiny app.
Generated using a Gaussian mixture model.
}
\keyword{datasets}
